[SPE-829] Clippy and fmt
