+++ src/s3.rs
@@ -14,7 +14,7 @@ use thiserror::Error;
 use super::{StorageClientError, StorageClientInterface};
 
 #[derive(Error, Debug)]
-pub enum S3ClientError {
+pub enum ClientError {
     #[error("GetObject Error: {0}")]
     GetObject(#[from] SdkError<GetObjectError>),
     #[error("PutObject Error: {0}")]
@@ -25,23 +25,23 @@ pub enum S3ClientError {
     General(String),
 }
 
-impl From<S3ClientError> for StorageClientError {
-    fn from(error: S3ClientError) -> Self {
+impl From<ClientError> for StorageClientError {
+    fn from(error: ClientError) -> Self {
         match error {
-            S3ClientError::GetObject(err) => StorageClientError::GetObject(err.to_string()),
-            S3ClientError::PutObject(err) => StorageClientError::PutObject(err.to_string()),
-            S3ClientError::DeleteObject(err) => StorageClientError::DeleteObject(err.to_string()),
-            S3ClientError::General(err) => StorageClientError::General(err),
+            ClientError::GetObject(err) => StorageClientError::GetObject(err.to_string()),
+            ClientError::PutObject(err) => StorageClientError::PutObject(err.to_string()),
+            ClientError::DeleteObject(err) => StorageClientError::DeleteObject(err.to_string()),
+            ClientError::General(err) => StorageClientError::General(err),
         }
     }
 }
 #[derive(Clone, Debug)]
-pub struct S3Client {
+pub struct StorageClient {
     bucket: String,
     client: s3::Client,
 }
 
-impl S3Client {
+impl StorageClient {
     #[allow(unused)]
     pub async fn new(bucket: String) -> Self {
         let region_provider = RegionProviderChain::default_provider().or_else("us-east-1");
@@ -52,7 +52,7 @@ impl S3Client {
 }
 
 #[async_trait]
-impl StorageClientInterface for S3Client {
+impl StorageClientInterface for StorageClient {
     async fn get_object(&self, key: String) -> Result<Option<String>, StorageClientError> {
         let object_res = self
             .client
@@ -76,12 +76,11 @@ impl StorageClientInterface for S3Client {
             .body
             .collect()
             .await
-            .map_err(|err| S3ClientError::General(err.to_string()))?
+            .map_err(|err| ClientError::General(err.to_string()))?
             .to_vec();
 
-        let body = String::from_utf8(body_bytes).map_err(|err| {
-            S3ClientError::General(format!("Failed to parse object body: {}", err))
-        })?;
+        let body = String::from_utf8(body_bytes)
+            .map_err(|err| ClientError::General(format!("Failed to parse object body: {err}")))?;
 
         Ok(Some(body))
     }
@@ -97,7 +96,7 @@ impl StorageClientInterface for S3Client {
             .body(ByteStream::from(body_bytes))
             .send()
             .await
-            .map_err(S3ClientError::PutObject)?;
+            .map_err(ClientError::PutObject)?;
 
         Ok(())
     }
@@ -110,7 +109,7 @@ impl StorageClientInterface for S3Client {
             .key(key)
             .send()
             .await
-            .map_err(S3ClientError::DeleteObject)?;
+            .map_err(ClientError::DeleteObject)?;
 
         Ok(())
     }
